MoveProfiling: outsource MoveType.
